+++ test_pdpyras.py
@@ -210,6 +210,22 @@ class APISessionTest(SessionTest):
         sess.raise_if_http_error = True
         self.assertRaises(pdpyras.PDClientError, list, sess.iter_all(weirdurl))
 
+    @patch.object(pdpyras.APISession, 'rpost')
+    @patch.object(pdpyras.APISession, 'iter_all')
+    def test_persist(self, iterator, creator):
+        user = {
+            "name": "User McUserson",
+            "email": "user@organization.com",
+            "type": "user"
+        }
+        iterator.return_value = iter([user])
+        sess = pdpyras.APISession('apiKey')
+        sess.persist('users', 'email', user)
+        creator.assert_not_called()
+        iterator.return_value = iter([])
+        sess.persist('users', 'email', user)
+        creator.assert_called_with('users', json=user)
+
     def test_postprocess(self):
         logger = MagicMock()
         response = Response(201, json.dumps({'key':'value'}), method='POST')
@@ -512,6 +528,23 @@ class APISessionTest(SessionTest):
             pdpyras.resource_envelope(do_http_things)(dummy_session,
                 '/incidents', json=incidents)
         )
+
+
+    @patch.object(pdpyras.APISession, 'put')
+    def test_resource_path(self, put_method):
+        sess = pdpyras.APISession('some-key')
+        resource_url = 'https://api.pagerduty.com/users/PSOMEUSR'
+        user = {
+            'id': 'PSOMEUSR',
+            'type': 'user',
+            'self': resource_url,
+            'name': 'User McUserson',
+            'email': 'user@organization.com'
+        }
+        put_method.return_value = Response(200, json.dumps({'user': user}),
+            method='PUT', url=resource_url)
+        sess.rput(user, json=user)
+        put_method.assert_called_with(resource_url, json={'user': user})
 
     @patch.object(pdpyras.APISession, 'get')
     def test_rget(self, get):