+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+**2020-06-26: Version 4.1**
+
+* Added new idempotent resource creator function, :attr:`APISession.persist`
+* Added the ability to use resource dictionaries (that have a ``self`` attribute) in place of URLs.
+
 **2020-02-04: Version 4.0**
 
 * Added support for using OAuth 2 access tokens to authenticate (`#23 <https://github.com/PagerDuty/pdpyras/issues/23>`_)